# docs/generated/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/v1/remote-assist/{id}/recording": {
            "post": {
                "description": "访客经 /api/v1/upload 上传录制文件后，把 key/时长/大小挂到对应协助会话；校验该协助归属当前访客的会话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "远程协助"
                ],
                "summary": "回写远程协助录制元数据（访客面）",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "协助会话 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "录制元数据",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.AttachRecordingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.RemoteAssistSession"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.AttachRecordingRequest": {
            "type": "object",
            "required": [
                "recording_key"
            ],
            "properties": {
                "recording_duration_ms": {
                    "type": "integer",
                    "example": 95000
                },
                "recording_key": {
                    "type": "string",
                    "example": "uploads/2026/09/rec-abc.webm"
                },
                "recording_mime": {
                    "type": "string",
                    "example": "video/webm"
                },
                "recording_size": {
                    "type": "integer",
                    "example": 1048576
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.RemoteAssistSession": {
            "type": "object",
            "properties": {
                "agent_user_id": {
                    "type": "integer"
                },
                "conversation_session_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "ended_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "recording_duration_ms": {
                    "type": "integer"
                },
                "recording_key": {
                    "description": "录制元数据：文件经既有 /api/v1/upload 上传，这里只落 key 与展示信息",
                    "type": "string"
                },
                "recording_mime": {
                    "type": "string"
                },
                "recording_size": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "description": "active|ended|failed",
                    "type": "string"
                },
                "tenant_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "workspace_id": {
                    "type": "string"
                }
            }
        },
...
    }
}